requirements.txt
fastapi==0.103.0
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
python-multipart==0.0.6
aiofiles==23.2.1
albumentations==1.3.1
streamlit==1.26.0
requests==2.31.0
requests-toolbelt==1.0.0
matplotlib==3.7.1
numpy==1.23.5
pandas==2.1.0
opencv-python==4.8.0.76
onnx==1.14.1
onnxruntime==1.15.1
segmentation_models_pytorch==0.3.0
tensorboard==2.14.0
torchmetrics==1.1.1
tqdm==4.66.1